Review comments fixed : Rearranged and reused code by creating a new Helper MenuItemTestHelper.java
